requirements.txt
gql
requests
python-dotenv
SQLAlchemy
pg8000